Move PtrRegister to lib/hstcalmemory.c and include/hstcal_memory.h
Resolves #236
