+++ ctegen2/wscript
@@ -14,6 +14,6 @@ def build(bld):
                     '../pkg/wfc3/calwf3/wf3cte/',
                     '../include/'
                     ],
-        use = ['CFITSIO', 'OPENMP'],
+        use = ['CFITSIO', 'OPENMP', 'hstcallib'],
         install_path = bld.env.INSTALL_LIB and '${PREFIX}/lib'
         )
