+++ pkg/acs/calacs/acscte/docte.c
@@ -10,6 +10,7 @@
 # include <time.h>
 # include <assert.h>
 
+#include "hstcal_memory.h"
 #include "hstcal.h"
 # include "hstio.h"
