+++ pkg/acs/calacs/acscte/wscript
@@ -14,9 +14,8 @@ def build(bld):
                   ],
         target = 'acscte.e',
         includes = ['.', '../../../../ctegen2'],
-        use = ['calacs', 'imphttab', 'ctegen2'] + bld.env.LOCAL_LIBS,
+        use = ['calacs', 'imphttab', 'ctegen2', 'hstcallib'] + bld.env.LOCAL_LIBS,
         lib = bld.env.EXTERNAL_LIBS,
         libpath = bld.env.LIBPATH,
         install_path = '${PREFIX}/bin'
         )
-
